requirements.txt
aiohttp==3.8.5
discord.py==2.3.2
psycopg2==2.9.9
python-dotenv==1.0.1
wheel==0.37.0
httpx==0.25.0
pykakasi==2.2.1
colorlog==6.8.2

